remove github workflow
All the github workflow-created packages I published to testpypi or pypi
installed but failed to work because they could not find the data
packages.  This is because .gitignore prevented the geopackage files
from being pushed to github!  Face-palm.  I don't want to commit the
data files to git, so I'm going to stick with building the packages
locally and uploading them using twine.  I don't need the github
workflow for this, so I'm removing it.